Notifier
Open program, enter URLs for websites (Without "http://" at the start),
and then the program will open them in your default browser whenever
those websites are updated, along with the first time the program checks
them. It will notify you when it checks a website and whether or not the
site was updated in a prompt window. Designed with
ebola-chan.co/index.php and ebola-chan.co/map.html in mind.

Separate multiple links with commas and no spaces.

Recommend for news websites, I'd guess, though I haven't tested it on
any as of yet. Except for ebola-chan.co, which it appears to work for.